pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "project-loom"
version = "0.1.0"
description = "Unified AI-ready workspace generator for multi-repo projects"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer[all]>=0.12",
    "questionary>=2.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.scripts]
loom = "loom.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["loom"]
